Cleanup tmp files support and droptables cfg add

+++ get_block_info.m
@@ -34,8 +34,13 @@ methods
                  ,'(id))');
             % obj.WriteLog(create_metric_table);
           
-            obj.drop_table();
+             if obj.cfg.DROP_TABLES
+                obj.WriteLog(sprintf("Dropping %s",obj.table_name))
+                obj.drop_table();
+                obj.WriteLog(sprintf("Dropped %s",obj.table_name))
+            end
             exec(obj.conn,create_metric_table);
+            
         end
               
         %drop table Striclty for debugging purposes
